Added deck create and list. Moved transactions to inner loop
