+++ src/models.rs
@@ -3,3 +3,9 @@ pub struct ListCard {
     pub front: String,
     pub back: String,
 }
+
+pub struct ListDeck {
+    pub id: i64,
+    pub name: String,
+    pub description: Option<String>,
+}
